# runtime_stack/condenser/docs/swagger.yaml
definitions:
...
  image.PullImageRequest:
    properties:
      arch:
        example: arm64
        type: string
      image:
        example: alpine:latest
        type: string
      os:
        example: linux
        type: string
    type: object
  image.RemoveImageRequest:
    properties:
      image:
        example: alpine:latest
        type: string
    type: object
...
  utils.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        description: success | fail
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/images:
    delete:
      consumes:
      - application/json
      description: remove image from local
      parameters:
      - description: Target Image
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/image.RemoveImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: remove image
      tags:
      - image
    get:
      description: get image list in local storage
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get image list
      tags:
      - image
    post:
      consumes:
      - application/json
      description: pull image from registry
      parameters:
      - description: Target Image
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/image.PullImageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: pull image
      tags:
      - image
  /v1/images/build:
    post:
      consumes:
      - application/x-tar
      description: build image from Dripfile-like archive (tar)
      parameters:
      - description: Target image tag (e.g. myapp:latest)
        in: query
        name: tag
        required: true
        type: string
      - description: 'Dripfile path in context (default: Dripfile)'
        in: query
        name: dripfile
        type: string
      - description: 'Bridge interface (default: raind0)'
        in: query
        name: network
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: build image
      tags:
      - image
  /v1/images/fs:
    get:
      description: get image filesystem usage
      parameters:
      - description: Target Image
        in: query
        name: image
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get image fs info
      tags:
      - image
  /v1/images/status:
    get:
      description: get image status details
      parameters:
      - description: Target Image
        in: query
        name: image
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get image status
      tags:
      - image
...
swagger: "2.0"